+++ net/http/retryable.go
@@ -8,9 +8,7 @@ import (
 	"fmt"
 	bytesext "github.com/go-playground/pkg/v5/bytes"
 	errorsext "github.com/go-playground/pkg/v5/errors"
-	ioext "github.com/go-playground/pkg/v5/io"
 	resultext "github.com/go-playground/pkg/v5/values/result"
-	"io"
 	"net/http"
 	"strconv"
 )
@@ -38,6 +36,15 @@ type ErrRetryableStatusCode struct {
 
 func (e ErrRetryableStatusCode) Error() string {
 	return fmt.Sprintf("retryable HTTP status code encountered: %d", e.Response.StatusCode)
+}
+
+// ErrUnexpectedResponse can be used to indicate an unexpected response was encountered as an error and provide access to the *http.Response.
+type ErrUnexpectedResponse struct {
+	Response *http.Response
+}
+
+func (e ErrUnexpectedResponse) Error() string {
+	return "unexpected response encountered"
 }
 
 // IsRetryableStatusCode returns if the provided status code is considered retryable.
@@ -106,9 +113,7 @@ func DoRetryable[T any](ctx context.Context, isRetryableFn errorsext.IsRetryable
 		defer resp.Body.Close()
 
 		if resp.StatusCode != expectedResponseCode {
-			b, _ := io.ReadAll(ioext.LimitReader(resp.Body, maxMemory))
-			err := fmt.Errorf("invalid response status code: %d body: %s", resp.StatusCode, string(b))
-			return resultext.Err[T, error](err)
+			return resultext.Err[T, error](ErrUnexpectedResponse{Response: resp})
 		}
 
 		data, err := DecodeResponse[T](resp, maxMemory)